pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Artyom <amordashov@gmail.com>"]
readme = "README.md"
repository = "https://github.com/yuvv2/mlopsHW"


[tool.poetry.dependencies]
python = "^3.11"
joblib = "^1.3.2"
pandas = "^2.2.0"
numpy = "^1.26.4"
dvc = "^3.47.0"
dvc-gdrive = "^3.0.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
mlflow = "^2.10.2"
scikit-learn = "1.4.0"
onnx = "^1.15.0"
skl2onnx = "^1.16.0"
onnxruntime = "^1.17.1"
tritonclient = {extras = ["http"], version = "^2.42.0"}
pre-commit = "^3.6.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[tool.black]
line-length = 90
target-version = ["py311"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B950']
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
    # file to define custom types
    'types.py: F401',
]
count = true

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "pyproject.toml"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
